discord_bot_v2/requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
Brotli==1.0.9
cchardet==2.1.7
cffi==1.15.0
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
nextcord==2.0.0a8
orjson==3.6.7
pycares==4.1.2
pycparser==2.21
PyNaCl==1.4.0
python-dotenv==0.19.2
six==1.16.0
wavelink==1.1.1
yarl==1.7.2
youtube-dl==2021.12.17
